Signals saved to hdf5 for each detector.

+++ libcore/detector_network.h
@@ -6,7 +6,7 @@
 
 /* The detector network used in the study. */
 typedef struct detector_network_s {
-	unsigned int num_detectors;
+	size_t num_detectors;
 	detector_t** detector;
 
 } detector_network_t;
